Move speech code to speech to text class.
